requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
sqlalchemy==2.0.41
asyncpg==0.30.0
alembic==1.15.2
toml==0.10.2
structlog==25.3.0

# Redis and caching
redis==6.1.0
aioredis==2.0.1
aiofiles==23.2.1

# Task processing
dramatiq==1.17.1

# Task scheduling
apscheduler==3.11.0

# Utilities
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.2.0
tenacity==9.1.2
python-dotenv==1.1.0

# Production dependencies
gunicorn==23.0.0
